Add google tracing
